requirements/local.txt
-r base.txt
ipdb
django-extensions
django-debug-toolbar
fabric
